Use float for pixelRatio

+++ include/mbgl/map/sprite.hpp
@@ -19,7 +19,7 @@ class FileSource;
 class SpritePosition {
 public:
     explicit SpritePosition() {}
-    explicit SpritePosition(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint8_t pixelRatio = 1);
+    explicit SpritePosition(uint16_t x, uint16_t y, uint16_t width, uint16_t height, float pixelRatio = 1.0f);
 
     operator bool() const {
         return !(width == 0 && height == 0 && x == 0 && y == 0);
@@ -27,7 +27,7 @@ public:
 
     uint16_t x = 0, y = 0;
     uint16_t width = 0, height = 0;
-    uint8_t pixelRatio = 1;
+    float pixelRatio = 1.0f;
 };
 
 class Sprite : public std::enable_shared_from_this<Sprite>, private util::noncopyable {
